+++ R/arg_between.R
@@ -37,9 +37,12 @@ arg_between <- function(x, range = c(0, 1), inclusive = TRUE,
                         .arg = rlang::caller_arg(x), .msg = NULL,
                         .call) {
 
-  arg_length(range, 2L, .call = rlang::current_env())
-  arg_logical(inclusive, .call = rlang::current_env())
-  arg_length(inclusive, 1:2, .call = rlang::current_env())
+  arg_length(range, 2L, .call = rlang::current_env()) |>
+    internal_arg()
+  arg_logical(inclusive, .call = rlang::current_env()) |>
+    internal_arg()
+  arg_length(inclusive, 1:2, .call = rlang::current_env()) |>
+    internal_arg()
 
   if (length(inclusive) == 1L) {
     inclusive <- inclusive[c(1L, 1L)]
@@ -60,7 +63,7 @@ arg_between <- function(x, range = c(0, 1), inclusive = TRUE,
       !safe_all(.lt_comp(x, range[2L]))) {
 
     if (is_not_null(.msg)) {
-      err(.msg)
+      err(.msg_eval(.msg))
     }
 
     each_element_of <- if (length(x) > 1L) "each element of"
@@ -98,13 +101,14 @@ arg_between <- function(x, range = c(0, 1), inclusive = TRUE,
 arg_gt <- function(x, bound = 0,
                    .arg = rlang::caller_arg(x), .msg = NULL,
                    .call) {
-  arg_length(bound, 1L, .call = rlang::current_env())
+  arg_length(bound, 1L, .call = rlang::current_env()) |>
+    internal_arg()
 
   if (!are_comparable(x, bound, `>`) ||
       !safe_all(x > bound)) {
 
     if (is_not_null(.msg)) {
-      err(.msg, .call = .call)
+      err(.msg_eval(.msg), .call = .call)
     }
 
     each_element_of <- if (length(x) > 1L) "each element of"
@@ -124,13 +128,14 @@ arg_gt <- function(x, bound = 0,
 arg_gte <- function(x, bound = 0,
                     .arg = rlang::caller_arg(x), .msg = NULL,
                     .call) {
-  arg_length(bound, 1L, .call = rlang::current_env())
+  arg_length(bound, 1L, .call = rlang::current_env()) |>
+    internal_arg()
 
   if (!are_comparable(x, bound, `>=`) ||
       !safe_all(x >= bound)) {
 
     if (is_not_null(.msg)) {
-      err(.msg, .call = .call)
+      err(.msg_eval(.msg), .call = .call)
     }
 
     each_element_of <- if (length(x) > 1L) "each element of"
@@ -150,13 +155,14 @@ arg_gte <- function(x, bound = 0,
 arg_lt <- function(x, bound = 0,
                    .arg = rlang::caller_arg(x), .msg = NULL,
                    .call) {
-  arg_length(bound, 1L, .call = rlang::current_env())
+  arg_length(bound, 1L, .call = rlang::current_env()) |>
+    internal_arg()
 
   if (!are_comparable(x, bound, `<`) ||
       !safe_all(x < bound)) {
 
     if (is_not_null(.msg)) {
-      err(.msg, .call = .call)
+      err(.msg_eval(.msg), .call = .call)
     }
 
     each_element_of <- if (length(x) > 1L) "each element of"
@@ -176,13 +182,14 @@ arg_lt <- function(x, bound = 0,
 arg_lte <- function(x, bound = 0,
                     .arg = rlang::caller_arg(x), .msg = NULL,
                     .call) {
-  arg_length(bound, 1L, .call = rlang::current_env())
+  arg_length(bound, 1L, .call = rlang::current_env()) |>
+    internal_arg()
 
   if (!are_comparable(x, bound, `<=`) ||
       !safe_all(x <= bound)) {
 
     if (is_not_null(.msg)) {
-      err(.msg, .call = .call)
+      err(.msg_eval(.msg), .call = .call)
     }
 
     each_element_of <- if (length(x) > 1L) "each element of"