+++ R/arg_character.R
@@ -25,7 +25,7 @@
 arg_character <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
                           .call) {
   if (!is.character(x)) {
-    err(.msg %or% "{.arg {(.arg)}} must be a character vector",
+    err(.msg_eval(.msg) %or% "{.arg {(.arg)}} must be a character vector",
         .call = .call)
   }
 }
@@ -35,7 +35,7 @@ arg_character <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
 arg_string <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
                        .call) {
   if (length(x) != 1L || !is.character(x) || anyNA(x)) {
-    err(.msg %or% "{.arg {(.arg)}} must be a string",
+    err(.msg_eval(.msg) %or% "{.arg {(.arg)}} must be a string",
         .call = .call)
   }
 }
@@ -45,7 +45,7 @@ arg_string <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
 arg_factor <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
                        .call) {
   if (!is.factor(x)) {
-    err(.msg %or% "{.arg {(.arg)}} must be a factor",
+    err(.msg_eval(.msg) %or% "{.arg {(.arg)}} must be a factor",
         .call = .call)
   }
 }