+++ R/arg_data.R
@@ -52,7 +52,7 @@
 arg_atomic <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
                        .call) {
   if (!is.atomic(x)) {
-    err(.msg %or% "{.arg {(.arg)}} must be an atomic vector",
+    err(.msg_eval(.msg) %or% "{.arg {(.arg)}} must be an atomic vector",
         .call = .call)
   }
 }
@@ -62,7 +62,7 @@ arg_atomic <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
 arg_vector <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
                        .call) {
   if (!is.atomic(x) || !is_null(dim(x))) {
-    err(.msg %or% "{.arg {(.arg)}} must be a vector",
+    err(.msg_eval(.msg) %or% "{.arg {(.arg)}} must be a vector",
         .call = .call)
   }
 }
@@ -72,15 +72,16 @@ arg_vector <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
 arg_list <- function(x, df_ok = FALSE,
                      .arg = rlang::caller_arg(x), .msg = NULL,
                      .call) {
-  arg_flag(df_ok, .call = rlang::current_env())
+  arg_flag(df_ok, .call = rlang::current_env()) |>
+    internal_arg()
 
   if (!df_ok && is.data.frame(x)) {
-    err(.msg %or% "{.arg {(.arg)}} must be a list and not a data frame",
+    err(.msg_eval(.msg) %or% "{.arg {(.arg)}} must be a list and not a data frame",
         .call = .call)
   }
 
   if (!is.list(x)) {
-    err(.msg %or% "{.arg {(.arg)}} must be a list",
+    err(.msg_eval(.msg) %or% "{.arg {(.arg)}} must be a list",
         .call = .call)
   }
 }
@@ -90,7 +91,7 @@ arg_list <- function(x, df_ok = FALSE,
 arg_data.frame <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
                            .call) {
   if (!is.data.frame(x)) {
-    err(.msg %or% "{.arg {(.arg)}} must be a data frame",
+    err(.msg_eval(.msg) %or% "{.arg {(.arg)}} must be a data frame",
         .call = .call)
   }
 }
@@ -100,7 +101,7 @@ arg_data.frame <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
 arg_matrix <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
                        .call) {
   if (!is.matrix(x)) {
-    err(.msg %or% "{.arg {(.arg)}} must be a matrix",
+    err(.msg_eval(.msg) %or% "{.arg {(.arg)}} must be a matrix",
         .call = .call)
   }
 }
@@ -110,7 +111,7 @@ arg_matrix <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
 arg_array <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
                       .call) {
   if (!is.array(x)) {
-    err(.msg %or% "{.arg {(.arg)}} must be an array",
+    err(.msg_eval(.msg) %or% "{.arg {(.arg)}} must be an array",
         .call = .call)
   }
 }
@@ -120,7 +121,7 @@ arg_array <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
 arg_data <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
                      .call) {
   if (!is.data.frame(x) && !is.matrix(x)) {
-    err(.msg %or% "{.arg {(.arg)}} must be a data frame or matrix",
+    err(.msg_eval(.msg) %or% "{.arg {(.arg)}} must be a data frame or matrix",
         .call = .call)
   }
 }
@@ -130,7 +131,7 @@ arg_data <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
 arg_env <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
                     .call) {
   if (!is.environment(x)) {
-    err(.msg %or% "{.arg {(.arg)}} must be an enviornment",
+    err(.msg_eval(.msg) %or% "{.arg {(.arg)}} must be an enviornment",
         .call = .call)
   }
 }