+++ R/arg_dots_supplied.R
@@ -25,7 +25,7 @@
 #' @export
 arg_dots_supplied <- function(..., .msg = NULL, .call) {
   if (...length() == 0L) {
-    err(.msg %or% "an argument must be supplied to {.arg ...}",
+    err(.msg_eval(.msg) %or% "an argument must be supplied to {.arg ...}",
         .call = .call)
   }
 }
@@ -34,7 +34,7 @@ arg_dots_supplied <- function(..., .msg = NULL, .call) {
 #' @rdname arg_dots_supplied
 arg_dots_not_supplied <- function(..., .msg = NULL, .call) {
   if (...length() != 0L) {
-    err(.msg %or% "no arguments may be supplied to {.arg ...}",
+    err(.msg_eval(.msg) %or% "no arguments may be supplied to {.arg ...}",
         .call = .call)
   }
 }