+++ R/arg_equal.R
@@ -39,10 +39,11 @@ arg_equal <- function(x, x2, ...,
                       .arg = rlang::caller_arg(x),
                       .arg2 = rlang::caller_arg(x2),
                       .msg = NULL, .call) {
-  arg_supplied(x2, .call = rlang::current_env())
+  arg_supplied(x2, .call = rlang::current_env()) |>
+    internal_arg()
 
   if (!isTRUE(all.equal(x, x2, ...))) {
-    err(.msg %or% "{.arg {(.arg)}} must be equal to {.arg {(.arg2)}}",
+    err(.msg_eval(.msg) %or% "{.arg {(.arg)}} must be equal to {.arg {(.arg2)}}",
         .call = .call)
   }
 }
@@ -53,10 +54,11 @@ arg_not_equal <- function(x, x2, ...,
                           .arg = rlang::caller_arg(x),
                           .arg2 = rlang::caller_arg(x2),
                           .msg = NULL, .call) {
-  arg_supplied(x2, .call = rlang::current_env())
+  arg_supplied(x2, .call = rlang::current_env()) |>
+    internal_arg()
 
   if (isTRUE(all.equal(x, x2, ...))) {
-    err(.msg %or% "{.arg {(.arg)}} must not be equal to {.arg {(.arg2)}}",
+    err(.msg_eval(.msg) %or% "{.arg {(.arg)}} must not be equal to {.arg {(.arg2)}}",
         .call = .call)
   }
 }