+++ R/arg_supplied.R
@@ -37,7 +37,8 @@ arg_supplied <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
 
     if (!rlang::is_symbol(arg_expr)) {
       err("{.arg x} must be an argument name",
-          .call = rlang::current_env())
+          .call = rlang::current_env()) |>
+        internal_arg()
     }
 
     err("an argument to {.arg {(.arg)}} must be supplied",